move control buttons top
