+++ website/src/plugins/rspressPluginPlayground/web/Playground/ControlPanel/ControlPanel.module.css
@@ -1,0 +1,17 @@
+.wrapper {
+  display: flex;
+  gap: 16px;
+  /* flex-wrap: wrap; */
+  align-items: center;
+  justify-content: space-between;
+
+  padding: 8px;
+}
+
+.section {
+  display: flex;
+  gap: 8px;
+  /* flex-wrap: wrap; */
+  align-items: center;
+  /* justify-content: space-between; */
+}
