--- website/src/plugins/rspressPluginPlayground/web/Playground/Preview/Preview.css
@@ -1,12 +1,0 @@
-.preview-actions {
-  position: absolute;
-  bottom: 8px;
-  right: 0;
-  margin-inline: 16px;
-
-  display: flex;
-  gap: 8px;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: flex-end;
-}
